Update UW Bay Area routing to not go to CA Online Intake
The entry in `vita_partners.yml` for United Way Bay Area actually
corresponded to the "Online Intake" group for the whole California
coalition.

Per partner support ticket #24957, we should update the routing such
that:
* "UWBA" source routing goes to UWBA (AK state routing remains with UWBA)
* "UWCA" source routing and CA state routing goes to CA Online Intake

I also reordered the YAML file so all the United Way of California
groups are next to each other.

One subtlety here is that our partner upsert behavior uses the partner's
group ID as a primary key, so the record for "United Way Bay Area" will
be renamed to "[United Way California] Online Intake", and a new record
for "United Way Bay Area" will be created. This might cause a bit of a
metrics incompatibility if we are looking at the partner name for "uwba"
source code routing. Ultimately, this seems pretty minor to me though.